% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEcellenImages.R
\name{CellenONEPlot}
\alias{CellenONEPlot}
\title{Create an Instance of a CellenONEPlot Panel}
\usage{
CellenONEPlot(...)
}
\arguments{
\item{...}{Arguments passed to the \code{initialize} method of the
\code{\link{CellenONEPlot-class}}. These can include initial values for
any of the slots defined in \code{CellenONEPlot-class} or its parent classes,
such as \code{ImageFile_Path}, \code{Background_Path}, \code{ImageDir}, etc.}
}
\value{
A \code{CellenONEPlot} object.
}
\description{
This function is the constructor for creating \code{\link{CellenONEPlot-class}} objects,
which represent an iSEE panel for visualizing CellenONE microscopy images.
}
\examples{
# Basic constructor
# CellenONEPlot()

# Constructor with custom slot values (if 'se' is a SummarizedExperiment)
# CellenONEPlot(ImageDir = "my_image_directory/", YAxisSampleName = colnames(se)[1])
}
\seealso{
\code{\link{CellenONEPlot-class}} for details on the class slots and methods.
}
\author{
Ed Emmott
}
