% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEHTMLMD.R
\docType{methods}
\name{HTMLpanel-class}
\alias{HTMLpanel-class}
\alias{HTMLpanel}
\alias{initialize,HTMLpanel-method}
\alias{.defineDataInterface,HTMLpanel-method}
\alias{.createObservers,HTMLpanel-method}
\alias{.hideInterface,HTMLpanel-method}
\alias{.fullName,HTMLpanel-method}
\alias{.panelColor,HTMLpanel-method}
\alias{.generateOutput,HTMLpanel-method}
\alias{.renderOutput,HTMLpanel-method}
\alias{.defineOutput,HTMLpanel-method}
\alias{.exportOutput,HTMLpanel-method}
\alias{.definePanelTour,HTMLpanel-method}
\title{The HTMLpanel Class}
\description{
The HTMLpanel class displays content from an iSEE app developer-specified Markdown or HTML file.
This is useful for including static documentation, for example an introduction or context for the
dataset.
}
\section{Slots}{

\describe{
\item{\code{FilePath}}{Character scalar specifying the path to the file (.md or.html) to be displayed.}
}}

\section{Slot overview}{

The following slots are relevant to the content:
\itemize{
\item \code{FilePath}, a string containing the path to a Markdown or HTML file.
The content of this file will be rendered and displayed inside the app.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\section{Constructor}{

\code{HTMLpanel(...)} creates an instance of an HTMLpanel class,
where any slot and its value can be passed to \code{...} as a named argument.
A \code{FilePath} argument \strong{must} be provided.
}

\seealso{
\linkS4class{Panel}
}
\author{
Ed Emmott
}
