% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEHTMLMD.R
\name{HTMLpanel}
\alias{HTMLpanel}
\title{Create an HTML Panel Instance}
\usage{
HTMLpanel(FilePath, ...)
}
\arguments{
\item{FilePath}{Character scalar specifying the path to the file
(.md or .html) to be displayed. This argument is mandatory.}

\item{...}{Additional arguments passed to the \code{initialize} method of
the parent \linkS4class{Panel} class, or to set other slots.}
}
\value{
An \code{HTMLpanel} object.
}
\description{
This function constructs an \code{HTMLpanel} object, an iSEE panel
for displaying content from a Markdown or HTML file.
}
\examples{
# panel <- HTMLpanel(FilePath = "path/to/your/file.md")
}
\seealso{
\code{\link{HTMLpanel-class}}
}
\author{
Ed Emmott
}
